pyproject.toml
[project]
name = "otel-instrumentation-nats"
version = "0.1.0"
description = "OpenTelemetry auto-instrumentation for nats-py"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opentelemetry-api ~= 1.27",
    "opentelemetry-instrumentation == 0.61b0",
    "opentelemetry-semantic-conventions == 0.61b0",
    "wrapt >= 1.0.0, < 3.0.0",
]

[project.optional-dependencies]
instruments = ["nats-py >= 2.0.0, < 3.0.0"]
test = [
    "pytest >= 7.0",
    "pytest-asyncio >= 0.21.0",
    "opentelemetry-sdk ~= 1.27",
    "nats-py >= 2.0.0, < 3.0.0",
]

[project.entry-points.opentelemetry_instrumentor]
nats = "otel_instrumentation_nats:NatsInstrumentor"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/otel_instrumentation_nats"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests requiring a running NATS server",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
